Player Data
Decided on the value of the player variables, I also made
"PlayerData.txt" to have another backup of them.

Changed the color of the default collider (png) to white (very useful as
you can now change the color in the editor, much easier to spot what is
what on screen now).

I also made some (very slight) changes in the player.cs,. Reason being I
found a problem with the way the speed was changed by the bady. The
speed itself never really changed, only the acceleration, (Meaning the
player could always get to the same maxspeed, but would only take much
longer to do so). This turn out to be very painful to play very quickly
(waiting to gain speed while moving pixel by pixel isn't the best way to
spend time imho). The changes I made address that and only that. The
effect of toxicity is now constant (before it was slowing the player
exponentially) and now reduce the maxSpeed, not the speed itself. This
make the player move slower without being near static while desperately
trying to gain speed, as the acceleration never changes.