Minor edition to LEVEL_ONE and metrics made
The metrics are finally here, only missing the amplitude that I have
absolutely no idea how to determine. The maximum value for length and
height between platform is there thought, as well as the character
proportion relative to the screen size. Made a visual guide in jpg and
edited the PlayerData.txt to add the new values.

Also changed the minimal speed value to 1.5.